feat(backend): rust ort integration
